feat: add unit test to threadpool
