+++ rust/src/pool/error.rs
@@ -1,6 +1,6 @@
 pub type Result<T> = std::result::Result<T, Error>;
 
-#[derive(Debug)]
+#[derive(Debug, PartialEq)]
 pub enum Error {
     InvalidPoolSize,
     ReceiverIsDropped,
